requirements.txt
flask==3.0.3
numpy==1.26.4
pillow==10.3.0
flask-cors==4.0.0
